Migrate Input implementation. #2
