+++ src/input.js
@@ -1,9 +1,9 @@
 import { createElement } from 'react'
 import { string } from 'prop-types'
-import createControl from '../create-control'
+import createControl from './create-control'
 
 export default createControl(({ control, ...props }) =>
-  createElement('input', { ...props, ...control })
+  createElement('input', { ...control, ...props })
 )({
   displayName: 'Input',
   propTypes: {
